内存访问崩溃因为第0帧没显著图

+++ common/frame.h
@@ -41,7 +41,7 @@ typedef struct x264_frame
         int     i_roi_x1;
         int     i_roi_x2;
     }roi;
-    int *salient;        /* salient data*/
+    uint8_t *salient;        /* salient data*/
     uint8_t *base;       /* Base pointer for all malloced data in this frame. */
     int     i_poc;
     int     i_delta_poc[2];
